+++ src/application-error-boundary/ApplicationErrorBoundary.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import StatusView from 'terra-status-view';
+import { injectIntl, intlShape } from 'react-intl';
 import logger from '../utils/logger';
 
 const propTypes = {
@@ -9,6 +10,12 @@ const propTypes = {
    * by these components during their render lifecycle will be caught by the ApplicationErrorBoundary.
    */
   children: PropTypes.node,
+
+  /**
+   * @private
+   * Intl object for translations.
+   */
+  intl: PropTypes.shape(intlShape),
 };
 
 /**
@@ -78,14 +85,16 @@ class ApplicationErrorBoundary extends React.Component {
   }
 
   render() {
-    const { children } = this.props;
+    const { children, intl } = this.props;
     const activeError = this.state.error || this.errorRef.current;
 
     if (activeError) {
+      const errorDetails = activeError.message.toString();
+      const errorText = intl.formatMessage({ id: 'terraApplication.errorBoundary.defaultErrorMessage' }, { errorDetails });
       return (
         <StatusView
           variant="error"
-          message={activeError.toString()}
+          message={errorText}
         />
       );
     }
@@ -96,4 +105,4 @@ class ApplicationErrorBoundary extends React.Component {
 
 ApplicationErrorBoundary.propTypes = propTypes;
 
-export default ApplicationErrorBoundary;
+export default injectIntl(ApplicationErrorBoundary);