+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Unreleased
 1.22.0 - (April 28, 2020)
 ------------------
 ### Changed
+* Updated errorBoundary to have a default error message when there is none provided.
 * Update terra-toolkit dev dependency to 6.0.0
 
 1.21.0 - (April 22, 2020)